works

+++ crates/core/src/globals.rs
@@ -7,7 +7,7 @@ use quickjs_wasm_rs::{JSContextRef, JSError, JSValue, JSValueRef};
 
 #[link(wasm_import_module = "codemod")]
 extern "C" {
-    fn __invokeHostFunc(func_idx: u64, a: u64) -> u64;
+    fn __invokeHostFunc(func_idx: u32, a: u64) -> u64;
 }
 
 static PRELUDE: &[u8] = include_bytes!("prelude/dist/index.js");
@@ -117,8 +117,8 @@ fn build_host_object(context: &JSContextRef) -> anyhow::Result<JSValueRef> {
         |_ctx: &JSContextRef, _this: JSValueRef, args: &[JSValueRef]| {
             let func_id = args.get(0).unwrap().as_i32_unchecked();
             let ptr = args.get(1).unwrap().as_u32_unchecked();
-            //let _result = invokeHostFunc(func_id as u64, ptr as u64) };
-            let result = unsafe { __invokeHostFunc(func_id as u64, ptr as u64) };
+            //let _result = invokeHostFunc(func_id as u32, ptr as u64) };
+            let result = unsafe { __invokeHostFunc(func_id as u32, ptr as u64) };
             extism_pdk::info!("{:#?}", &result);
             Ok(JSValue::Int(result as i32))
         },
